DERBY-3882: Expensive cursor name lookup in network server
Statements executed via the network client are always given cursor
names on the server, even if the user has not set a cursor name. When
a statement has a cursor name, the embedded driver will check on each
execution that there is no other statement in the same connection with
the same cursor name and an open result set. To perform this check,
the list of activations in the connection is traversed and each cursor
name is compared with the cursor name of the statement to be executed.

If the number of open statements in the connection is high, which is
very likely if ClientConnectionPoolDataSource with the JDBC statement
cache is used, traversing the list of activations and performing
string comparisons may become expensive.

This patch attempts to make this operation cheaper without performing
a full rewrite and use a new data structure. It exploits that the most
common implementations of java.lang.String cache the hash code, so
calling hashCode() on the cursor names in the list will simply read
the value from an int field after warm-up. By checking if the hash
codes of the cursor names match first, we can avoid many of the string
comparisons because we know that strings with different hash codes
cannot be equal.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@818807 13f79535-47bb-0310-9956-ffa450edef68