DERBY-4326
Rather than quiting out of ClientThread for InterruptedException and IOException, we should first check if the exceptions aree raised because of server being shutdown. If yes, then we should quit out of ClientThread. Otherwise, we should log the exception and close the client socket that got the exception and go back to the infinite loop waiting for next client connection.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@823659 13f79535-47bb-0310-9956-ffa450edef68