DERBY-3002: Add partial support for ROLLUP variant of GROUP BY clause
This change adds an initial implementation of support for the ROLLUP
style of GROUP BY. Specifying ROLLUP causes Derby to perform multiple
nested levels of grouping in a single pass over the input data, returning
the results in a single result set.

Most of the changes were made to GroupedAggregateResultSet, which is
enhanced to be able to perform multiple levels of grouping in a single
pass over the (sorted) input data. There are also changes to the SQL grammar
to allow the new syntax to be specified, and changes to GroupByNode to
pass the rollup specification through to the execution engine.

Several tests were added to the OLAPTest test suite.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@824966 13f79535-47bb-0310-9956-ffa450edef68