DERBY-3634 Cannot use row_number() in ORDER BY clause
Patch derby-3634-newimpl-4. This reimplements the ROW_NUMBER
functionality after it was moved in svn 820483 and 820494. The patch
makes ROW_NUMBER usable in all places it should be usable according to
the standard (modulo the current Derby SQL subset).  So, DERBY-3634 is
also fixed with this patch. The ordering problem seen in DERBY-4069 is
also fixed. More test cases have been to OLAPTest for ROW_NUMBER
usage.  The patch also adds some amount of generalizations to pave way
for other window functionality, but this needs more work.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@829155 13f79535-47bb-0310-9956-ffa450edef68