DERBY-4120: derbyclient.jar is not a complete OSGi bundle
OSGi bundles are JAR files with special information in their manifest to
enable the OSGi class loader to provide dynamic load/unload functionality.

This change adds some additional information to the manifest of
derbyclient.jar to support its use as an OSGi bundle. In particular,
it marks the org.apache.derby.jdbc package as exported, since that package
contains the ClientDriver and ClientDataSource classes which applications
are likely to use directly.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@830545 13f79535-47bb-0310-9956-ffa450edef68