DERBY-4420: NullPointerException with INSERT INTO ... from EXCEPT/INTERSECT
The failing code in ResultSetNode.setTableConstructorTypes() was meant
to handle the case where the node represented a table constructor (aka
VALUES clause). UnionNode already had an override to make it a no-op
unless it actually represented a multi-row VALUES clause that had been
rewritten to a union of single-row VALUES clauses.

Since a VALUES clause is never rewritten to EXCEPT or INTERSECT, the
correct handling is to make setTableConstructorTypes() a no-op in
IntersectOrExceptNode. Rather than adding an empty override in
IntersectOrExceptNode, the code was moved from
ResultSetNode.setTableConstructorTypes() to
RowResultSetNode.setTableConstructorTypes(), and the default
implementation in ResultSetNode was left empty.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@832379 13f79535-47bb-0310-9956-ffa450edef68