DERBY-4124: Remove double-checked-locking from EmbeddedDataSource.findDriver
This change ensures that EmbeddedDataSource.findDriver always synchronizes
its access to the 'driver' member variable. The previous code performed
unsynchronized access to the field, and then, depending on the results of
the unsynchronized access, performed synchronized access.

This double-checked-locking idiom is unsafe and incorrect, and this change
replaces it with a simpler always-synchronized access.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@833888 13f79535-47bb-0310-9956-ffa450edef68