DERBY-4450 GROUP BY in an IN-subquery inside HAVING clause whose select list is subset of group by columns, gives NPE
Patch DERBY-4450b + Knut's simplification of the autocommit call in GroupByTest#testDerby4450.

This solves the problem seen in this issue, which was a regression
from DERBY-681. The crux of the problem is that a PRN is added in the
result set tree without adjusting a higher level reference so that
wrong code is generated. The solution here is to reuse the result
column list in the inserted PRN, so that reference from above will
point correctly even after the PRN insertion (more details in JIRA). 

Backported from trunk as

svn merge -c 882732 https://svn.apache.org/repos/asf/db/derby/code/trunk

plus a manual fixup.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.4@882965 13f79535-47bb-0310-9956-ffa450edef68