DERBY-3225: Include demo README information in demo/programs/readme.html
This change updates the README information for the demo/programs area of
the binary distribution to have additional information about the "localcal",
"scores", and "vtis" demos. These demo programs were added in DERBY-1936,
DERBY-1993, and DERBY-3129, respectively.

This change also includes the build.xml files for the "localcal" and "scores"
demos into the binary distribution.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@883571 13f79535-47bb-0310-9956-ffa450edef68