DERBY-4451 ArrayIndexOutOfBoundsException or ASSERT FAILED when inserting generated columns out of order
This patch fixes this issue as well as DERBY-4448, since they share
the same underlying problem: the former way of checking for illegal
override of generated columns in the presence of an explicit target
column list failed to look below the top level UnionNode in a table
constructor.

This specialized treatment for the case of an explicit target column
list has been removed for INSERT (it was shared with logic for
UPDATE), so checking is now done by the code for the case of no
explicit targte column list.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@884163 13f79535-47bb-0310-9956-ffa450edef68