DERBY-151 Thread termination -> XSDG after operation is 'complete'
This patch catches the exception seen by RAFContainer4
readFull/Write/full when the embedding application thread sees an
interrupt, and asks the user to inspect the app to find the cause.
Before this patch, the symptoms were puzzling to the user, suggesting
that the disk might be full.

The database level error seen is XSDG9:
"Derby thread received an interrupt during a disk I/O operation, please check your application for the source of the interrupt. "

A new store test, Derby151Test, checks this behavior.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@886831 13f79535-47bb-0310-9956-ffa450edef68