DERBY-4387 Infinite loop in PredicateList.joinClauseTransitiveClosure()
This patch, derby-4387, fixes a corner case of the closure code in
which the outer and the inner table is the same. This led to adding
the same, new predicate twice to the same table, causing the infinite
loop. The patch also adds a test case, which, without the fix, shows
the infinite loop.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@887246 13f79535-47bb-0310-9956-ffa450edef68