DERBY-4400: Document the process of producing Maven 2 artifacts for Derby
Added a utility class that obtains the Derby version from the jars that will
be put into the artifacts. It also does a few sanity checks of the jars, for
instance checking if they are all there.
Updated the README and changed/added some comments in settings.xml.

Patch file: derby-4400-1c-SetDerbyVersion.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@888385 13f79535-47bb-0310-9956-ffa450edef68