DERBY-4405 Transformation to inner join not performed for certain three-way joins
Patch derby-4405-2, which extends the rewriting optimization of outer
join to inner join in the presence of a compound inner table which
itself a join. The existing analysis of null intolerant predicates,
which is a condition for the rewrite, failed to look into inner tables
that were not a simple base table.

The patch adds new test cases to verify that rewrite now takes place
for such cases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@891015 13f79535-47bb-0310-9956-ffa450edef68