DERBY-4496 Column list size mismatch with ORDER BY in INSERT statement
Fixed this issue which was due to missed occurences of RCL.size() that
now (with ORDER BY in subqueries) need to be generalized to
RCL.visibleSize to be able to ignore synthesized columns. Added a new
test case for the failing repro. Also wired the OrderByInSubqueries
test into the regression suite.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@896388 13f79535-47bb-0310-9956-ffa450edef68