DERBY-4397 Allow ORDER BY in subqueries
Patch derby-4397-sortavoidance-a, which makes sort avoidance work in
the simple case of a subquery with order by on a base table which has
a suitable index. The fix avoids the result columns assumed by the
order by list from changing after they have been bound, to include a
level of VCNs, which is not correct (this happened as a side effect of
copying result columns from the subquery up to the FromSubquery node).

A new test case the verify sort avoidance in this case has been added.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@899588 13f79535-47bb-0310-9956-ffa450edef68