DERBY-4512: Avoid unnecessary lookup in transaction table when adding transaction
The transaction table should not contain a transaction with the same
id as the one that's added with TransactionTable.add(). Therefore, put
the new transaction in the table unconditionally instead of checking
first whether there is an existing entry to reuse. This saves one
lookup in the transaction table per transaction.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@900714 13f79535-47bb-0310-9956-ffa450edef68