derby-4477 Selecting / projecting a column whose value is represented by a stream more than once fails
Patch derby-4477-partial-2. This patch clones streamable columns in
occurence 2..n in ProjectRestrictResultset if they occur more than
once in the select list. It also adds the three repro test cases from
DERBY-3645, DERBY-3646 and DERBY-2349 to BLOBTest. The patch is not
complete and needs to be revisted when cloning of store streams is
implemented. Currenly the cloning occurs via materialization and this
will exhaust memory when lobs are large.  See FIXME in
ProjectRestrictResultSet.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@902857 13f79535-47bb-0310-9956-ffa450edef68