DERBY-4477 Selecting / projecting a column whose value is represented by a stream more than once fails
Patch derby-4477-lowmem-2, which adds test cases to check that lobs
are not materialized when large, for the use cases covered by this
issue. The test cases are added to the lowmem suite, which is not not
part of the regular suites.All. This commit is preparatory in that the
lobs are still small, so these changes should be revisited to change
their sizes when the cloning handles materialization properly, cf
DERBY-3650 and DERBY-4520.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@904538 13f79535-47bb-0310-9956-ffa450edef68