DERBY-4477 Selecting / projecting a column whose value is represented by a stream more than once fails
Patch derby-4477-useCloning: Uses the new stream cloning referred to
earlier as a coming follow-up for this issue: derby-4477-useCloning.

It also bumps the lob size in the lowmem suite's test cases added for
this issue in ClobMemTest and BlobMemTest, which will run out of
memory without this cloning used by this patch.

Note:

The "length-materialization optimization" (code that was commented out
earlier, and removed by the latest patch) hasn't been implemented
yet. However, we think it is better placed inside the various
cloneValue-methods and hope to get to it at a later time.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@908563 13f79535-47bb-0310-9956-ffa450edef68