DERBY-3729: Improve client-side messages when file or disk fills up.
This change is intended to improve the error messages which are returned
by the Derby client JDBC calls when error message XSDG1 occurs. This
particular error message may arise when the disk fills up, or it may
arise due to a quote or filesystem limit. For example, on FAT32 file
systems, an individual file is limited to 4GB and when that limit is
reached, error XSDG1 occurs.

Delivering error information back to the client in this case is tricky,
because the XSDG1 error is a database-severity message, meaning that the
database is closed, the connection is closed, and the session is ended,
so the information must be send to the client as part of the Command
Check Response Message itself.

This change: adds additional text to the XSDG1 and J123 messages, and
enhances the client's handling of the CMDCHKRM to look for possible
error information in the SQLCARD.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@911955 13f79535-47bb-0310-9956-ffa450edef68