DERBY-4341: Building with a different CLASSPATH causes runmessagecheck to fail.
This change was contributed by John Storta Jr. (js at storta dot net)

The problem was that running the MessageBundleTest as an Ant task made
it hard for us to control the exact CLASSPATH that was used, because
Ant tasks always are affected by the overall CLASSPATH that is being
used for the Ant process.

The change converts MessageBundleTest from an Ant task to a standalone Java
program, and changes the runmessagecheck target in the top-level build.xml
to run MessageBundleTest in a separate JVM with fork=true so that the
CLASSPATH setting can be exactly controlled.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@911999 13f79535-47bb-0310-9956-ffa450edef68