DERBY-1482
We do not have many tests for triggers on tables with LOB columns. Checking in few tests with this checkin. The tests are disabled at this point because they run out of memory when run with limited heap because of streaming of the LOB columns whether they are needed or not. The test is part of lowmem suite which will automatically take care of running the tests with limited heap which is 16M for that suite at this point.

More information on the test can be found in the comments in the test and also in the jira.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@915177 13f79535-47bb-0310-9956-ffa450edef68