DERBY-4574: derby.jar needs Bundle-ManifestVersion header for OSGi
The derby.jar has all the bundle metadata you would expect for it to
be an OSGi bundle, except for the Bundle-ManifestVersion header.
As a result it is not a valid R4 bundle, but will be treated as an R3 bundle.

Looking at the metadata provided it looks as if the bundle should be
an R4 bundle. Adding the Bundle-ManifestVersion will make it a proper R4 bundle. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@918152 13f79535-47bb-0310-9956-ffa450edef68