DERBY-4322  intermittent failure in runtimeinfo.
Contributed by Myrna van Lunteren.

The theory is that this failure is caused by a problem in the previous test DerbyNetAutostart which launches a process to start the server and can be left pinging if the check for the server in the launching process, does not allow enough time for the server to come up.  The pings increment the session number as seen by runtimeinfo.  So the changes are to DerbyNetAutoStart and are to

1) Make sure the server launching process retries enough times for the launched process comes up.
2) Make sure if something does go wrong, the launched process is destroyed



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@925377 13f79535-47bb-0310-9956-ffa450edef68