DERBY-4582: Timestamps inserted with GMT calendar are 1 hour later when subsequently read with GMT calendar (Server Mode Only)
Partial fix. This check-in makes the client do the conversion to the
requested time zone correctly, and the correct timestamps will be sent
to the server. There's still some incorrect handling of the values on
the server, as well as when the client reads them back from the
server, so the results may still end up being wrong.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@935868 13f79535-47bb-0310-9956-ffa450edef68