DERBY-4582: Timestamps inserted with GMT calendar are 1 hour later when subsequently read with GMT calendar (Server Mode Only)
Partial fix. This fix makes the server use a GMT calendar when it
reads a date, time or timestamp from the client. This way, the values
won't be changed if they refer to a point in time that does not exist
in the local time zone due to switch from/to daylight saving time (GMT
doesn't observe daylight saving time and therefore doesn't have these
holes in the range of valid values).


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@936218 13f79535-47bb-0310-9956-ffa450edef68