DERBY-2017: Client driver can insert and commit partial data when a LOB stream throws IOException or does not match the specified length
Made the client tell the server (if supported) if the EXTDTA transfer was
successful or not, by appending a Derby-specific status byte to the user data.
If the transfer was unsuccessful, an exception will be thrown on the server
side even if the data was padded by the client to avoid a protocol exception.
There are three types of errors that can happen on the client when reading
the user stream:
  a) the stream is shorter than specified
  b) the stream is longer than specified
  c) an IOException is thrown when reading the stream

The fix consists of logic to allow the client and the server to determine if
the status byte will be sent, code to append the status byte on the client, and
code on the server to parse / read the status byte.
Some code in DRDAConnThread was also refactored (mostly readAndSetExtParam).

Patch file: derby-2017-3c-fix.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@936950 13f79535-47bb-0310-9956-ffa450edef68