DERBY-4582: Timestamps inserted with GMT calendar are 1 hour later when subsequently read with GMT calendar (Server Mode Only)
Partial fix. This patch makes the server read time, date and timestamp
values from ResultSets and CallableStatements using the getter methods
that take a Calendar object. A GMT calendar is used, so that times
that are not possible to represent in the local time zone (like the
"lost hour" when switching to daylight saving time) are not silently
modified. The GMT calendar is also used when formatting the value
before it's sent to the client.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@937170 13f79535-47bb-0310-9956-ffa450edef68