DERBY-4314 With derby client setTrasactionIsolation executes and commits even if isolation has not changed.
Contributed by Lily Wei (lilywei at yahoo dot com) and Kristian Waagan
setTransactionIsolation will no longer commit.
Piggybacking of the isolation level was implemented to avoid a performance impact.
There is specialized logic for XA Resume and Join to make sure the client isolation is in sync with the server.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@940620 13f79535-47bb-0310-9956-ffa450edef68