DERBY-4667 BootLockTest.testBootLock() sometimes fails with connection refused.
Move the vm logic for Derby151Test outside of the fixture so network server won't start if no tests are run.  This may not correct the problem and even if it does may be a fragile fix, but it giving it a try to see if the prior test Derby151Test is interfering if no fixtures are run but network server is stopped and started.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@950818 13f79535-47bb-0310-9956-ffa450edef68