DERBY-4693: RENAME COLUMN loses IDENTITY attributes
This change modifies the RENAME COLUMN logic so that it preserves the
autoincrement properties of the column when renaming it.  Since RENAME
COLUMN more-or-less drops and re-adds the column, we need to ensure
that when the column is re-added, if it is an Autoincrement column, 
we set the CREATE_AUTOINCREMENT flag that the parser sets when
initially creating the table. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@954344 13f79535-47bb-0310-9956-ffa450edef68