DERBY-1482 (Update triggers on tables with blob columns stream blobs into memory even when the blobs are not referenced/accessed.)
Reduced the size of LOBs in the test to 50MB since junit-lowmem gets run with 16MB and hence no need to read unnecessarily larger LOB(the test was testing with 300MB lobs). 

Also, some subset of tests are relying on DriverManager which is not available with J2ME. I have disabled those subset of tests to not run under J2ME framework.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@957785 13f79535-47bb-0310-9956-ffa450edef68