DERBY-2785: ij "describe" command cannot describe table named "run"
This fix was contributed by Eranda Sooriyabandara (070468D at gmail dot com)

The IJ tool's parser was not flexible enough in its handling of
identifiers, and so if you have a table named RUN, the command

ij> describe run;

was treating RUN as a keyword, rather than realizing it was the name of
the table. By changing the IJ parser's identifier() production to accept
keywords, the command is now accepted and behaves as desired.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@958431 13f79535-47bb-0310-9956-ffa450edef68