DERBY-4241: Improve transition from read-only to writable Clob representation
When a store stream Clob is going to be modified, it will be written out to the temporary area of Derby and represented as a TemporaryClob.
The transfer of the data is done in a sub-optimal manner for two reasons;
 o for transfer of the complete Clob, the copy method operates on the byte level and we're not able to save the character length.
 o for transfer of parts of the Clob (i.e. truncation), we have to first decode the UTF-8 encoding to find the byte count and then transfer the same bytes.

This fix addresses both issues by improving the length caching and by adding
a method that counts the charactes "on-the-fly" while copying raw bytes.

Patch file: derby-4241-2b-utf8AwareCopy.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@958522 13f79535-47bb-0310-9956-ffa450edef68