DERBY-4694: Build breaks on OS X due to JDK classpath issues.
Enabled JAR inspection for Mac OS X instead of using hard-coded values.
The motivation for the change is that Apple has started symlinking the
directories for JDK 1.4 and JDK 1.5 to JDK 1.6. This made the build fail on
OS X (only recent fresh installs, where the symlinking is used).

Patch file: derby-4694-1b.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@958555 13f79535-47bb-0310-9956-ffa450edef68