DERBY-4577 An expanding update fails with an nospc.U error
Fixes problem on Overflow pages populated by long rows where the an expanding
update of the row would fail with a nospc.U error.  The system always expects
updates to succeed.  In the case of an expanding update that finds not enough
room there should always at least be enough room reserved/used in the row to
do an update that changes the row to just an overflow row pointer with no data,
moving the data to another page.

The problem was that on overflow pages not enough room was being reserved to
handle a worst case row with a overflow pointer.  This fix reserves the worst
case space on overflow pages.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@959027 13f79535-47bb-0310-9956-ffa450edef68