DERBY-4698 Simple query with HAVING clause crashes with NullPointerException
backporting fix with svn #956234 from trunk to 10.5 branch.

Patch DERBY-4698-2. The case of column references in HAVING clauses
being wrong after JOIN flattening was initially solved by
DERBY-3880. That solution was partial in that it can sometimes happen
too late.

This patch changes the fix-up of column references in a having clause
after join flattening to the same point in time as that of other
column references that need fix-up after the flattening (rcl, column
references in join predicates and group by clauses). Thus the fixup is
moved from the modifyaccesspath phase to the preprocess phase. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@959697 13f79535-47bb-0310-9956-ffa450edef68