DERBY-4479: RENAME TABLE needs to invalidate any cached CREATE TABLE statement
backporting fix svn #909176 from trunk to 10.5 branch.

The problem involves a missing dependency between the CREATE TABLE statement
and the table that it is creating. For other types of statements, the
dependency of the statement on the table is generally registered during
compilation.

However, this isn't quite so easy with the CREATE TABLE statement, because
the table descriptor doesn't exist during compilation, since the table isn't
created until the statement is actually executed.

This change registers the dependency at execution time. At the very end of
CreateTableConstantAction.executeConstantAction, there is
some new code added to register a dependency from the CREATE TABLE
statement to the table we have just finished creating. 





git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@960105 13f79535-47bb-0310-9956-ffa450edef68