DERBY-4244: ALTER TABLE ASSERT in ADD COLUMN with autocommit off
backported svn #795459 from trunk to 10.5 branch.  Also backported the
AlterTableTest.java from trunk so that test for this issue could be 
run.

This patch was contributed by Eranda Sooriyabandara (070468D at gmail dot com)

An ALTER TABLE ... ADD COLUMN statement which specifies to add a column to
a table may also include any column constraints which apply to that column.
For example, to add a column to a table and at the same time indicate that
the column should be the primary key of the table, this statement is used:

   alter table t0 add column c2 int not null default 0 primary key;

If the table already contains a primary key, this statement fails, as the
table cannot have more than one primary key. However, the check for the
duplicate primary key was occurring at execution time, *after* the column
had already been added to the table. Then, during that same transaction,
subsequent ALTER TABLE statements affecting that table failed with internal
ASSERT messages because the data dictionary data structures were inconsistent.

This patch adds logic to the compilation processing of the ALTER TABLE
statement to check for a duplicate PRIMARY KEY constraint at compile
time, before the ALTER TABLE statement has begun execution. This way, when
such a duplicate constraint is rejected, the rejection occurs prior to
execution of the table, allowing for more straightforward cleanup.

The patch also re-enables a portion of the AlterTableTest, which had been
disabled due to this problem. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@960222 13f79535-47bb-0310-9956-ffa450edef68