DERBY-4425: NPE with INSERT .. SELECT .. UNION and generated expressions
backporting change #831319 from trunk to 10.5.

This patch adds a test which confirms that DERBY-4425 is fixed. The code
change which resolved DERBY-4425 was performed as part of fixing DERBY-4419,
since the same code change fixes both symptoms, so this is a test-only patch.

subversion revision 831304 contains the DERBY-4419 code change. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@960617 13f79535-47bb-0310-9956-ffa450edef68