DERBY-4155 jdbcapi/XATest.java doesn't seem to be running anywhere
backporting the current state of XATest.java and XAtestUtil.java from trunk
to 10.3 branch.   Disabled the test for DERBY-4731, to be enabled later when
that is backported.  It looks like all of the changes for the rest of this
issue were already backported to 10.3 as part of the backport of DERBY-4310.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@962687 13f79535-47bb-0310-9956-ffa450edef68